Added drawer navigation Added Drawer header data frm repository Added Token in NetworkConstants.kt Fixed padding on textInputField Separated Authentication and Customer data repositories Added NetworkInterceptor.kt to add header in Api requests

+++ app/src/main/java/com/devsinc/bws/model/FeaturedClass.kt
@@ -6,7 +6,7 @@ data class FeaturedClass(
     val class_image: String,
     val class_location: String,
     val class_name: String,
-    val class_rating: Int,
+    val class_rating: Float,
     val class_slag: String,
     val class_tag: List<String>
 )
